+++ plutip-server/Api/Handlers.hs
@@ -5,10 +5,11 @@ module Api.Handlers (
 
 import Cardano.Api (serialiseToCBOR)
 import Cardano.Launcher.Node (nodeSocketFile)
-import Cardano.Wallet.Shelley.Launch.Cluster (RunningNode (RunningNode))
+import Test.Plutip.Tools.CardanoApi qualified as Tools
+
+import Control.Concurrent.MVar (isEmptyMVar, putMVar, tryTakeMVar)
 import Control.Concurrent (threadDelay)
 import Control.Concurrent.Async (race)
-import Control.Concurrent.MVar (isEmptyMVar, putMVar, tryTakeMVar)
 import Control.Monad (unless)
 import Control.Monad.Except (runExceptT, throwError)
 import Control.Monad.Extra (unlessM)
@@ -16,19 +17,29 @@ import Control.Monad.IO.Class (liftIO)
 import Control.Monad.Reader (ReaderT, ask, asks)
 import Data.ByteString.Base16 qualified as Base16
 import Data.Default (def)
-import Data.Either (fromRight)
 import Data.Foldable (for_)
+import Data.Either (fromRight)
 import Data.List.Extra (firstJust)
 import Data.Text.Encoding qualified as Text
 import Data.Traversable (for)
 import System.Directory (doesFileExist)
 import System.FilePath (replaceFileName)
-import Test.Plutip.Config (chainIndexPort, relayNodeLogs)
+import Test.Plutip.Config (
+  ChainIndexMode (NotNeeded),
+  PlutipConfig (extraConfig),
+  chainIndexMode,
+  relayNodeLogs,
+ )
 import Test.Plutip.Internal.BotPlutusInterface.Setup (keysDir)
-import Test.Plutip.Internal.BotPlutusInterface.Wallet (BpiWallet (signKey), addSomeWallet)
+import Test.Plutip.Internal.BotPlutusInterface.Wallet (
+  BpiWallet (signKey),
+  addSomeWallet,
+  cardanoMainnetAddress,
+ )
+import Test.Plutip.Internal.Cluster (RunningNode (RunningNode))
+import Test.Plutip.Internal.Cluster.Extra.Types (ExtraConfig (ExtraConfig, ecSlotLength))
 import Test.Plutip.Internal.LocalCluster (startCluster, stopCluster)
-import Test.Plutip.Internal.Types (ClusterEnv (runningNode))
-import Test.Plutip.LocalCluster (cardanoMainnetAddress)
+import Test.Plutip.Internal.Types (ClusterEnv (plutipConf, runningNode))
 import Test.Plutip.Tools.CardanoApi (AwaitWalletFundedError (AwaitingCapiError, AwaitingTimeoutError), awaitWalletFunded)
 import Types (
   AppM,
@@ -48,7 +59,12 @@ import Types (
   Lovelace (unLovelace),
   PrivateKey,
   ServerOptions (ServerOptions, nodeLogs),
-  StartClusterRequest (StartClusterRequest, keysToGenerate),
+  StartClusterRequest (
+    StartClusterRequest,
+    epochSize,
+    keysToGenerate,
+    slotLength
+  ),
   StartClusterResponse (
     ClusterStartupFailure,
     ClusterStartupSuccess
@@ -61,7 +77,7 @@ import UnliftIO.Exception (throwString)
 startClusterHandler :: ServerOptions -> StartClusterRequest -> AppM StartClusterResponse
 startClusterHandler
   ServerOptions {nodeLogs}
-  StartClusterRequest {keysToGenerate} = interpret $ do
+  StartClusterRequest {slotLength, epochSize, keysToGenerate} = interpret $ do
     -- Check that lovelace amounts are positive
     for_ keysToGenerate $ \lovelaceAmounts -> do
       for_ lovelaceAmounts $ \lovelaces -> do
@@ -70,10 +86,12 @@ startClusterHandler
     statusMVar <- asks status
     isClusterDown <- liftIO $ isEmptyMVar statusMVar
     unless isClusterDown $ throwError ClusterIsRunningAlready
-    let cfg = def {relayNodeLogs = nodeLogs, chainIndexPort = Nothing}
+    let extraConf = ExtraConfig slotLength epochSize
+        cfg = def {relayNodeLogs = nodeLogs, chainIndexMode = NotNeeded, extraConfig = extraConf}
+
     (statusTVar, (clusterEnv, wallets)) <- liftIO $ startCluster cfg setup
     liftIO $ putMVar statusMVar statusTVar
-    res <- liftIO $ race (threadDelay 2_000_000) $ waitForFundingTxs clusterEnv wallets
+    res <- liftIO $ race (threadDelay 2_000_000) $ waitForFundingTxs clusterEnv wallets extraConf
     -- throw Exception for cardano-cli errors.
     -- Ignore wait timeout error - return from this handler doesn't guarantee funded wallets immedietely.
     maybe (return ()) throwString $ fromRight Nothing res
@@ -91,16 +109,16 @@ startClusterHandler
     where
       setup :: ReaderT ClusterEnv IO (ClusterEnv, [BpiWallet])
       setup = do
+        env <- ask
         wallets <- do
           for keysToGenerate $ \lovelaceAmounts -> do
             addSomeWallet (fromInteger . unLovelace <$> lovelaceAmounts)
-        env <- ask
         return (env, wallets)
 
-      -- wait for confirmation of funding txs, throw the first error if there's any
-      waitForFundingTxs clusterEnv wallets = do
+       -- wait for confirmation of funding txs, throw the first error if there's any
+      waitForFundingTxs clusterEnv wallets extraConfig = do
         res <- for wallets $ \w ->
-          awaitWalletFunded clusterEnv (cardanoMainnetAddress w)
+          awaitWalletFunded clusterEnv (cardanoMainnetAddress w) extraConfig
         return $
           firstJust
             ( \case
@@ -114,7 +132,6 @@ startClusterHandler
       getNodeConfigFile =
         -- assumption is that node.config lies in the same directory as node.socket
         flip replaceFileName "node.config" . getNodeSocketFile
-
       getWalletPrivateKey :: BpiWallet -> PrivateKey
       getWalletPrivateKey = Text.decodeUtf8 . Base16.encode . serialiseToCBOR . signKey
       interpret = fmap (either ClusterStartupFailure id) . runExceptT
@@ -127,4 +144,4 @@ stopClusterHandler StopClusterRequest = do
     Nothing -> pure $ StopClusterFailure "Cluster is not running"
     Just statusTVar -> do
       liftIO $ stopCluster statusTVar
-      pure $ StopClusterSuccess
+      pure StopClusterSuccess