+++ src/Test/Plutip/Predicate.hs
@@ -29,8 +29,19 @@ import Data.Map (Map)
 import Data.Map qualified as Map
 import Ledger (ExBudget (ExBudget), ExCPU (ExCPU), ExMemory (ExMemory), TxId, Value)
 import PlutusCore.Evaluation.Machine.ExMemory (CostingInteger)
+import Prettyprinter (
+  Doc,
+  align,
+  defaultLayoutOptions,
+  indent,
+  layoutPretty,
+  viaShow,
+  vsep,
+  (<+>),
+ )
+import Prettyprinter.Render.String (renderString)
 import Test.Plutip.Internal.Types (
-  ExecutionResult (contractState, outcome),
+  ExecutionResult (ExecutionResult, contractState, outcome),
   FailureReason (CaughtException, ContractExecutionError),
   budgets,
   isSuccessful,
@@ -83,8 +94,22 @@ shouldSucceed =
   Predicate
     "Contract should succeed"
     "Contract should fail"
-    (mappend "But it didn't.\nResult: " . show)
+    (mappend "But it didn't.\nResult: " . renderString . layoutPretty defaultLayoutOptions . prettyExecutionResult)
     isSuccessful
+
+-- | Pretty print ExecutionResult hiding budget stats and logs.
+prettyExecutionResult :: (Show e, Show w, Show a) => ExecutionResult e w a -> Doc ann
+prettyExecutionResult ExecutionResult {outcome, contractState} =
+  vsep
+    [ "Execution result {"
+    , indent 4 $
+        align $
+          vsep
+            [ "outcome:" <+> viaShow outcome
+            , "final state: " <+> viaShow contractState
+            ]
+    , "}"
+    ]
 
 -- | Check that Contract didn't succeed.
 --
@@ -309,8 +334,8 @@ budgetsFitUnder (Limit sCpu sMem) (Limit pCpu pMem) =
 
       -- "filter" `TxBudget` by removing only script or minting budgets
       -- that fit limits
-      findNonFitting b =
-        filterBudget (Prelude.not . fits sCpu sMem) (Prelude.not . fits pCpu pMem) b
+      findNonFitting =
+        filterBudget (Prelude.not . fits sCpu sMem) (Prelude.not . fits pCpu pMem)
 
       fits cpuLimit memLimit (ExBudget cpu' mem') =
         cpu' <= cpuLimit && mem' <= memLimit