+++ src/Test/Plutip/Tools/CardanoApi.hs
@@ -1,28 +1,42 @@
+{-# LANGUAGE PartialTypeSignatures #-}
+{-# OPTIONS_GHC -Wno-partial-type-signatures #-}
+
 -- | Helpers based on Cardano.Api (do not use `cardano-cli` executable)
 module Test.Plutip.Tools.CardanoApi (
   currentBlock,
   utxosAtAddress,
   queryProtocolParams,
   queryTip,
+  awaitAddressFunded,
   awaitWalletFunded,
+  plutusValueFromAddress,
+  CardanoApiError,
   AwaitWalletFundedError (AwaitingCapiError, AwaitingTimeoutError),
 ) where
 
 import Cardano.Api qualified as C
-import Cardano.Api.Shelley (ProtocolParameters)
+import Cardano.Api.Shelley (ProtocolParameters, TxOut (TxOut), UTxO (UTxO, unUTxO), txOutValueToValue)
 import Cardano.Launcher.Node (nodeSocketFile)
 import Cardano.Slotting.Slot (WithOrigin)
-import Cardano.Wallet.Shelley.Launch.Cluster (RunningNode (RunningNode))
-import Control.Arrow (right)
+import Test.Plutip.Internal.Cluster (RunningNode (RunningNode))
+
 import Control.Exception (Exception)
-import Control.Retry (constantDelay, limitRetries, retrying)
+import Control.Arrow (right)
+import Control.Monad.Catch (MonadMask)
+import Control.Monad.IO.Class (MonadIO (liftIO))
+import Control.Monad.Reader (MonadReader (ask), ReaderT)
+import Control.Retry (constantDelay, limitRetries, recoverAll, retrying)
 import Data.Either (fromRight)
-import Data.Map qualified as M
+import Data.Map qualified as Map
 import Data.Set qualified as Set
+import Data.Time (NominalDiffTime, nominalDiffTimeToSeconds)
 import GHC.Generics (Generic)
+import Ledger (Value)
+import Ledger.Tx.CardanoAPI (fromCardanoValue)
 import Ouroboros.Consensus.HardFork.Combinator.AcrossEras (EraMismatch)
-import Ouroboros.Network.Protocol.LocalStateQuery.Type (AcquireFailure)
 import Test.Plutip.Internal.Types (ClusterEnv (runningNode))
+import Test.Plutip.Internal.Cluster.Extra.Types (ExtraConfig (ecSlotLength))
+import UnliftIO (throwString)
 
 newtype CardanoApiError
   = SomeError String
@@ -31,7 +45,7 @@ newtype CardanoApiError
 instance Exception CardanoApiError
 
 -- | Get current block using `Cardano.Api` library
-currentBlock :: ClusterEnv -> IO (Either AcquireFailure (WithOrigin C.BlockNo))
+currentBlock :: ClusterEnv -> IO (Either _ (WithOrigin C.BlockNo))
 currentBlock (runningNode -> rn) = do
   let query = C.QueryChainBlockNo
       info = connectionInfo rn
@@ -78,6 +92,43 @@ flattenQueryResult = \case
   Right (Right res) -> Right res
   err -> Left $ SomeError (show err)
 
+-- | Waits till specified address is funded using `CardanoApi` query.
+-- Performs 60 tries with `retryDelay` seconds between tries.
+awaitAddressFunded ::
+  (MonadIO m, MonadMask m) =>
+  C.AddressAny ->
+  NominalDiffTime ->
+  ReaderT ClusterEnv m ()
+awaitAddressFunded addr retryDelay = do
+  cEnv <- ask
+  recoverAll policy $ \_ -> do
+    utxo <- liftIO $ utxosAtAddress cEnv addr
+    checkUtxo utxo
+  where
+    delay = truncate $ nominalDiffTimeToSeconds retryDelay * 1000000
+    policy = constantDelay delay <> limitRetries 60
+
+    checkUtxo = \case
+      Left e ->
+        throwString $
+          "Failed to get UTxO from address via cardano API query: "
+            <> show e
+      Right (UTxO utxo')
+        | Map.null utxo' ->
+          throwString "No UTxOs returned by cardano API query for address"
+      _ -> pure ()
+
+-- | Get total `Value` of all UTxOs at address.
+plutusValueFromAddress ::
+  ClusterEnv ->
+  C.AddressAny ->
+  IO (Either CardanoApiError Value)
+plutusValueFromAddress cEnv addr = do
+  let getValues = mconcat . fmap extract . (Map.elems . unUTxO)
+      extract (TxOut _ txoV _ _) = fromCardanoValue $ txOutValueToValue txoV
+  res <- utxosAtAddress cEnv addr
+  return $ getValues <$> res
+
 data AwaitWalletFundedError
   = AwaitingCapiError CardanoApiError
   | AwaitingTimeoutError
@@ -87,20 +138,19 @@ instance Show AwaitWalletFundedError where
   show AwaitingTimeoutError = "Awaiting funding transaction timed out."
 
 -- | Waits till specified address is funded using cardano-node query.
--- Performs 20 tries with 0.2 seconds between tries, which should be a sane default.
--- Waits till there's any utxos at an address - works for us as funds will be send with tx per address.
+-- Performs 60 tries with `retryDelay` seconds between tries.
 awaitWalletFunded ::
   ClusterEnv ->
   C.AddressAny ->
+   ExtraConfig ->
   IO (Either AwaitWalletFundedError ())
-awaitWalletFunded cenv addr = toErrorMsg <$> retrying policy checkResponse action
+awaitWalletFunded cenv addr extraConfig = toErrorMsg <$> retrying policy checkResponse action
   where
-    -- With current defaults the slot length is 0.2s and block gets produced about every second slot.
-    -- We are expected to wait 0.4s, waiting 4s we are almost guaranteed (p>0.9999)
-    delay = 200_000 -- in microseconds, 0.2s.
-    policy = constantDelay delay <> limitRetries 20
+    retryDelay = ecSlotLength $ extraConfig
+    delay = truncate $ nominalDiffTimeToSeconds retryDelay * 1000000
+    policy = constantDelay delay <> limitRetries 60
 
-    action _ = right (M.null . C.unUTxO) <$> utxosAtAddress cenv addr
+    action _ = right (Map.null . C.unUTxO) <$> utxosAtAddress cenv addr
 
     checkResponse _ = return . fromRight False